requirements.txt
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
blinker==1.9.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1749686179973/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1749229904370/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1752493286596/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560565029/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click==8.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1744220033875/work/conda-src
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1742219570693/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1751548120229/work
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
Flask==3.1.2
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1728841357330/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302889955/work
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1746515861348/work/libmambapy
MarkupSafe==3.0.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1753546308902/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1747339660894/work
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1732588447334/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1749498106507/work
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1749479969459/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724470051/work
setuptools==80.9.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1739009763/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
Werkzeug==3.1.3
wheel==0.45.1
zipstream-new==1.1.8
zstandard==0.23.0
